+++ synthelion/nlp/__init__.py
@@ -2,5 +2,9 @@ from synthelion.nlp.text_splitter import TextSplitter
 from synthelion.nlp.sentence_detector import SentenceDetector
 from synthelion.nlp.summarizer import TfIdfSummarizer
 from synthelion.nlp.text_rank import TextRankSummarizer
+from synthelion.nlp.topic_segmenter import TopicSegmenter, TopicSegment
 
-__all__ = ["TextSplitter", "SentenceDetector", "TfIdfSummarizer", "TextRankSummarizer"]
+__all__ = [
+    "TextSplitter", "SentenceDetector", "TfIdfSummarizer", "TextRankSummarizer",
+    "TopicSegmenter", "TopicSegment",
+]